+++ speechsauce/src/feature.rs
@@ -26,10 +26,10 @@ use ndrustfft::{nddct2, DctHandler};
 pub(crate) fn filterbanks(
     num_filter: usize,
     coefficients: usize,
-    sampling_freq: f64,
-    low_freq: Option<f64>,
-    high_freq: Option<f64>,
-) -> Array2<f64> {
+    sampling_freq: f32,
+    low_freq: Option<f32>,
+    high_freq: Option<f32>,
+) -> Array2<f32> {
     //TODO: compare to https://pytorch.org/audio/main/_modules/torchaudio/functional/functional.html#melscale_fbanks
     let high_freq = high_freq.unwrap_or(sampling_freq / 2.0);
     let low_freq = low_freq.unwrap_or(300.0);
@@ -43,7 +43,7 @@ pub(crate) fn filterbanks(
     // converting the upper and lower frequencies to Mels.
     // num_filter + 2 is because for num_filter filterbanks we need
     // num_filter+2 point.
-    let mels = Array1::<f64>::linspace(
+    let mels = Array1::<f32>::linspace(
         frequency_to_mel(low_freq),
         frequency_to_mel(high_freq),
         num_filter + 2,
@@ -56,7 +56,7 @@ pub(crate) fn filterbanks(
     // exact points calculated above should be extracted.
     //  So we should round those frequencies to the closest FFT bin.
     let freq_index = mel_arr_to_frequency(mels)
-        .map(|x| ((coefficients + 1) as f64 * x / sampling_freq) as usize);
+        .map(|x| ((coefficients + 1) as f32 * x / sampling_freq) as usize);
 
     // Initial definition
     let mut filterbank = Array2::zeros([num_filter, coefficients]);
@@ -67,12 +67,12 @@ pub(crate) fn filterbanks(
         let middle = freq_index[i + 1];
         let right = freq_index[i + 2];
 
-        let z = Array1::<f64>::linspace(left as f64, right as f64, right - left + 1);
+        let z = Array1::<f32>::linspace(left as f32, right as f32, right - left + 1);
 
         {
-            let mut s: ArrayViewMut1<f64> = filterbank.slice_mut(s![i, left..right + 1]);
+            let mut s: ArrayViewMut1<f32> = filterbank.slice_mut(s![i, left..right + 1]);
 
-            s.assign(&triangle(z, left as f64, middle as f64, right as f64));
+            s.assign(&triangle(z, left as f32, middle as f32, right as f32));
         }
     }
     filterbank
@@ -82,16 +82,16 @@ pub(crate) fn filterbanks(
 ///     Args:
 ///          signal : the audio signal from which to compute features.
 ///              Should be an N x 1 array
-pub fn mfcc(signal: ArrayView1<f64>, speech_config: &SpeechConfig) -> Array2<f64> {
+pub fn mfcc(signal: ArrayView1<f32>, speech_config: &SpeechConfig) -> Array2<f32> {
     let (mut feature, energy) = mfe(signal, &speech_config);
 
     if feature.is_empty() {
-        return Array::<f64, _>::zeros((0_usize, speech_config.num_cepstral));
+        return Array::<f32, _>::zeros((0_usize, speech_config.num_cepstral));
     }
     feature = feature.log();
     //feature second axis equal to num_filters
     let feature_axis_len = feature.shape()[1];
-    let mut transformed_feature = Array2::<f64>::zeros(feature.raw_dim());
+    let mut transformed_feature = Array2::<f32>::zeros(feature.raw_dim());
     //link to og code:
     //https://github.com/astorfi/speechpy/blob/2.4/speechpy/feature.py#L147
     //function dct docs:
@@ -103,13 +103,13 @@ pub fn mfcc(signal: ArrayView1<f64>, speech_config: &SpeechConfig) -> Array2<f64
     //4. norm="ortho": when used with dct type 2, applies a scaling factor
 
     //need to switch to rustfft, provide len of last axis specifically
-    let mut dct_handler: DctHandler<f64> = DctHandler::new(feature_axis_len);
+    let mut dct_handler: DctHandler<f32> = DctHandler::new(feature_axis_len);
 
     //need to check how to specify axis of transformation
     nddct2(&feature, &mut transformed_feature, &mut dct_handler, 1);
     //NOTE: may be able to remove the if/else by processing first element separately
     //https://docs.scipy.org/doc/scipy/reference/generated/scipy.fftpack.dct.html#:~:text=2N)-,If%20norm%3D%27ortho%27%2C%20y%5Bk%5D%20is%20multiplied%20by%20a%20scaling%20factor%20f,-f%3D%7B
-    let n = transformed_feature.len() as f64;
+    let n = transformed_feature.len() as f32;
     //orthonormalized the transformed axis
     transformed_feature[[0, 0]] *= 1. / (4. * n).sqrt();
     transformed_feature
@@ -134,13 +134,13 @@ pub fn mfcc(signal: ArrayView1<f64>, speech_config: &SpeechConfig) -> Array2<f64
 }
 //TODO: https://pytorch.org/audio/main/_modules/torchaudio/transforms/_transforms.html#MelSpectrogram
 //https://github.com/librosa/librosa/blob/c800e74f6a6ec5c27e0fa978d7355943cce04359/librosa/feature/spectral.py#LL2021C5-L2021C5
-fn mel_spectrogram(signal: ArrayView1<f64>, speech_config: &SpeechConfig) {
+fn mel_spectrogram(signal: ArrayView1<f32>, speech_config: &SpeechConfig) {
     //ndarray::einsum
     todo!("implement mel_spectrogram")
 }
 ///a helper function that is passed to stack_frames from mfe
-fn _f_it(x: usize) -> Array2<f64> {
-    Array2::<f64>::ones((x, 1))
+fn _f_it(x: usize) -> Array2<f32> {
+    Array2::<f32>::ones((x, 1))
 }
 
 /// Compute Mel-filterbank energy features from an audio signal.
@@ -163,7 +163,7 @@ fn _f_it(x: usize) -> Array2<f64> {
 ///    Returns:
 ///         array: features - the energy of fiterbank of size num_frames x num_filters.
 ///         The energy of each frame: num_frames x 1
-pub fn mfe(signal: ArrayView1<f64>, speech_config: &SpeechConfig) -> (Array2<f64>, Array1<f64>) {
+pub fn mfe(signal: ArrayView1<f32>, speech_config: &SpeechConfig) -> (Array2<f32>, Array1<f32>) {
     //
     // Stack frames
     let frames = stack_frames(
@@ -198,7 +198,7 @@ pub fn mfe(signal: ArrayView1<f64>, speech_config: &SpeechConfig) -> (Array2<f64
 ///         speech_config: the configuration for the speech processing functions
 ///    Returns:
 ///         array: Features - The log energy of fiterbank of size num_frames x num_filters frame_log_energies. The log energy of each frame num_frames x 1
-fn lmfe(signal: ArrayView1<f64>, speech_config: &SpeechConfig) -> Array2<f64> {
+fn lmfe(signal: ArrayView1<f32>, speech_config: &SpeechConfig) -> Array2<f32> {
     let (feature, _) = mfe(signal, speech_config);
     feature.log()
 }
@@ -209,7 +209,7 @@ fn lmfe(signal: ArrayView1<f64>, speech_config: &SpeechConfig) -> Array2<f64> {
 ///     feature : The feature vector which its size is: N x M
 /// Return:
 ///     array: The feature cube vector which contains the static, first and second derivative features of size: N x M x 3
-fn extract_derivative_feature(feature: Array2<f64>) -> Array3<f64> {
+fn extract_derivative_feature(feature: Array2<f32>) -> Array3<f32> {
     let first_derivative_feature = crate::processing::derivative_extraction(&feature, 2);
     let second_derivative_feature =
         crate::processing::derivative_extraction(&first_derivative_feature, 2);
